refactor: Apply reviewer feedback for code modifications
